created TestScreen and added Buttons to move across all of the screens
