+++ app/src/main/java/com/example/clasetrabajo/ui/screens/MainMenuScreen.kt
@@ -1,10 +1,14 @@
 package com.example.clasetrabajo.ui.screens
 
+
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 
@@ -13,8 +17,15 @@ import androidx.navigation.NavHostController
 fun MainMenuScreen(navController: NavHostController){
     Column(
         modifier = Modifier
+            .padding(10.dp)
             .fillMaxSize()
     ){
       Text("Main Menu")
+        Button(
+            onClick = {navController.navigate("homeScreen")}
+        )
+        {
+            Text("Go to Home Screen")
+        }
     }
 }
